This fragment
module Bug where
class Eq a => Coll c a where
  fn :: (a -> Bool) -> a -> c a

bug a = fn (\y -> y == y) a

cause some versions of Classic Hugs to barf. It was caused by a
missing comparison, checking for a precondition.

    /* the h==h1 test is just an optimization, and I'm not
       sure it will work with IPs, so I'm being conservative
       and commenting it out - JL */
    /* We need h == h1, as it appears to be a precondition
       for checking if the predicate are the same - AG */
    if (h==h1 && samePred(pi1,o1,pi,o))
	return e;

If the comment above is correct, then this 'fix' might
break something in IP.