Applied Sigbjorn's bugfix for the parser.
(Without this fix, you can't load lib/exts/Int.hs.
I guess it should be added to the mythical test suite.
Which reminds me... what happened to the test suite?
We used to have one.  I think it was in hugs/src/tests.
A (admittedly superficial) search failed to reveal it.)