Hugs98 implementation of Time.
The implementation is broken since it sits on top of libc & its 
myopic view of time as being (UNIX Epoch) +/- ~68 years. i.e,. it
follows in the footstep of what was done on the GHC side. Can do
better though, but this implementation will have to do for now.

Regression tests to follow soon.