Fix UI niggle - when loading the module
  module Foo where { import BrassMoney; }

Hugs will complain if a file corresponding to BrassMoney isn't
in scope. So, the user spots the obvious typo in the Foo module,
changing it to

  module Foo where { import BrassMonkey; }

However, doing a :reload still caused BrassMoney to be searched for,
with equally miserable results. The only way to purge the bogus
module from the stack is to do a :load.

This commit fixes this, giving :reload the desired behaviour.