Added many of the ffi types (FunPtr, Ptr, etc) to Hugs and wired the {Int,Word}{8,16,32,64} types more firmly into Prelude/compiler
We still lack ForeignPtr. What is trivial for GHC with its threads and
all the rest of it is a scary proposition for the rest of us.
(Specifically, I don't know if it is safe to start evaluating
finalizers at all places that the GC might be invoked.  Maybe the eval
function could be changed to consult a list of finalizers now and then
- that would be somewhat gross but safe.)

The 64 bit arithmetic types lack any operations.

Next step: update the code generation part of ffi, implement the
currently unsupported foreign import decl and see if I can port the
foreign export dynamic (or whatever it is called) sparc code from GHCi
into Hugs.