First of a series of commits of the new ffi code.  Existing uses of GreenCard-generated code should still work (and seems to do so).
These are enough of the libraries to run some tests.

Some of them (notably CTypes.hs and Storable.hs) are very incomplete
_and_ should (but aren't) be configured with autoconf and should (but
isn't) be built from the sources in the hierarchial library archive.
I'd love it if someone were to tackle either of these problems before
I get to them.