The previous commit changed how hugs +G works.
It used to be that the generated file was compiled in the directory
the file came from.  That is, we executed:

   cd $(dirof ${file}) ; cc $(notdir ${file})

I couldn't get this to work on Win98 so I changed the design.
We no longer do the cc.

The result is that any files, includes, etc. specified on the cc line
are relative to the current working directory not the directory the
ffi file is found in.

This commit fixes MkInstall/in to assume this.

Coming soon: fix the testsuit to accomodate this change.


[I also reran autoconf which I don't have on my Win32 box.]