More Sun4 cc changes.
cc starts spewing warning messages when confronted with pre-ANSI
function prototypes and declarations which involve char, short, etc.
Such functions are, unavoidably, part of the ffi code.  Fix is to give
in and admit that ffi code is _not_ K&R compatible.  This was probably
true before, converting it all to ANSI-style function prototypes and
definitions merely confirms it.