First attempt at improving modularity of the Hugs codebase.
This commit splits up and factors out the numerable tasks that
the main hugs.c file had become burdened with over the years.
The following new modules have been introduced:

  * evaluator.{c,h} -- evaluator engine
  * errors.{c,h}    -- error handling + output
  * goal.{c,h}      -- handling progress towards interp goals.
  * observe.{c,h}   -- observation specific code
                       (Note: support is only available in interactive builds.)
  * opts.{c,h}      -- parsing and representing interpreter options.
  * strutil.{c,h}   -- string utilities (not too much in here at present, but
                       will grow.)
  * HugsAPI.h       -- the public header file for the HugsAPI; server.h is now
  		       assumed to be internal to the Hugs sources.
  * server.c        -- now a separate module; no longer literally includes hugs.c
  * hugs.c          -- now only responsible for handling the interactive UI and
                       all its commands.
  * winhugs/winhugs.c -- separate module; no longer literally included by hugs.c
  		         Code that supports a separate evaluator thread has been
			 moved to here (from hugs.c)

This was all brought about by wanting to expose the HugsAPI / server
interface to external code (invoked via the FFI), and also re-use
some of the server.c code to provide an extended version of the
HugsAPI. What seemed like an innocent little change at first
turned out not to be!

[Untested on autoconf-based plats; true to form, I've no doubt mucked up
Makefile.in some.]