Change the representation of the strict ST monad to
	newtype ST s a = ST (forall r. (a -> r) -> r)

which is a generalization of the new representation of IO monad:

	newtype IO a = IO ((a -> IOResult) -> IOResult)

This simplifies conversion between these monads, and enables them to
share monad and state primitives, so that stmonad.c is no longer required
(except for a couple of primitives to support Hugs.Memo, which have
moved to iomonad.c).

The lazy ST monad has the same representation as before, but is now
defined in Hugs.LazyST, with conversions to and from the strict ST monad.
We now use the fptools/libraries definitions of the operations on this
monad, using the corresponding operations on the strict monad and these
conversions.  The resulting interface matches that of GHC.