Make searching of the current directory for modules configurable.
Currently, when resolving module names to file names, Hugs searches the
current directory before those listed in hugsPath.  With this change,
only hugsPath is used, but the default setting of hugsPath includes
the current directory at the start, preserving the current behaviour.
However this may now be overridden.

Note that this relates only to resolving module names -- file names
don't use the path.  Also, when a file name is specified, the directory
containing it is searched first for any modules imported by that file
(and other modules found in that directory).