Change the behaviour of the :browse command (if "all" not specified): formerly listed names defined in the module, now lists exported names.
Since the libraries now use imports much more, with the old behaviour
things like ":b List" listed no names, which is both useless and confusing
to beginners.  Now it lists the exports of List (which is also much
closer to the output before we switched to the hierarchical libraries).

Also added a few rev's to keep the export lists in the order they
were specified.