from Neil Mitchell: repair interact and getChar for WinHugs.
The implementation in this version is superior to that in the old version
of WinHugs, including that the stop button now stops immediately (not
after one char is pressed) and the status bar changes to advise the user
that input is required.

The approach uses a Content semaphore to pause the calculating thread
when no input is available. Like all threading code, this is not
particularly fun to write or easy to understand.