Added a shiny new unified darcs-all script
This is a mildly tested new version of a darcs-all script which should work
with GHC, Hugs, and nhc98. It is not yet a full 100% replacement of the
existing scripts, but it will soon be. Apart from the fact that it is not tied
to a single Haskell implementation, its main new features are that it uses a
single more powerful package description file and offers a --release option to
retrieve a given set of package versions from Hackage.

To do:

 * Handle more general URLs in the package list

 * Add a mode to print out the list of packages (for Makefiles, scripts, etc.)

 * --complete/--partial: What should be the default?