Add type signatures to numeric constants printed in corehugs mode.
For proper generation of Yhc Core, it is necessary to distinguish between Int and Integer,
and between Float and Double constants. This patch modifies the print routines in src/output.h
so that in corehugs mode (that is, with +C), numeric constants are printed like this:
(42 :: Int). This does not affect the interactive Hugs.
In addition, when Hugs is configured with --enable-extcore, the +C option is disabled
for interactive Hugs (it does not make much sense anyway to run Hugs with +C interactively).